# storage/openapi.yaml
openapi: 3.0.3
info:
  title: statekit storage API
  version: 0.3.0
  description: >
    Query API for stored statekit state documents, organized in three data
    layers. L1 (/state/summary, /state/targets) is the constantly polled
    current view without data payloads. L2 (/state/targets/{key},
    /state/states/{identity}) is the full current detail fetched on
    drill-down. L3 (/state/states/{identity}/timeline, /state/timeline,
    /state/timeline/bucket, incidents) is historical and explicitly bounded.
    Every layer GET sets a strong ETag and honors If-None-Match with 304, so
    an idle fleet polls as one small summary body plus 304s.
paths:
  /state/summary:
    get:
      summary: Fleet rollup (L1)
      description: >
        Worst status, state and target status counts, fleet hash, and the
        latest observation time. fleet_hash is the ETag; it is the hash of
        the sorted (target key, material_hash) pairs and changes only when
        some target's material content changes.
      responses:
        '200':
          description: Fleet summary
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetSummary'
        '304':
          $ref: '#/components/responses/NotModified'
  /state/targets:
    get:
      summary: List target summaries (L1)
      description: >
        One summary per target with flat per-state headers and no data
        payloads. The ETag is the fleet hash; each summary's material_hash is
        the ETag of its L2 detail document.
      responses:
        '200':
          description: Target summaries
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TargetSummary'
        '304':
          $ref: '#/components/responses/NotModified'
  /state/targets/{key}:
    get:
      summary: Full detail for one target (L2)
      description: >
        The target summary plus full state details including data. The ETag
        is the target's material_hash from the L1 summary.
      parameters:
        - name: key
          in: path
          required: true
          description: Target key as returned by /state/targets, URL-encoded.
          schema:
            type: string
      responses:
        '200':
          description: Target detail
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetDetail'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
  /state/states/{identity}:
    get:
      summary: Full detail for one state (L2)
      description: One state's detail including data and child identities.
      parameters:
        - name: identity
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: State detail
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateDetail'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
  /state/states/{identity}/timeline:
    get:
      summary: Transition history for one state (L3)
      description: >
        The state's bounded transition ring, newest first. Transitions are
        keyed by identity + changed_at + status; scrapes of an unchanged
        state add nothing.
      parameters:
        - name: identity
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: State timeline
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateTimeline'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
  /state/timeline:
    get:
      summary: Bucketed triggering-state counts (L3 charting store)
      description: >
        Draws the stacked "which states were triggering over time" chart and
        the sparkline directly; one small array, no client-side aggregation.
        Counts are distinct triggering states per status per bucket.
      parameters:
        - name: scope
          in: query
          description: '"fleet" (default) or "target:{key}".'
          schema:
            type: string
        - name: window
          in: query
          description: Go duration, for example 24h or 90m. Defaults to 24h.
          schema:
            type: string
        - name: buckets
          in: query
          description: Number of display buckets. Defaults to 64.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Bucketed counts
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BucketCounts'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
  /state/timeline/bucket:
    get:
      summary: Triggering states in one chart bucket (L3 charting store)
      description: >
        Tooltip/hover detail; the triggering states recorded in the native
        bucket containing t. Labels are captured at write time, so entries
        resolve without joining the current layers.
      parameters:
        - name: scope
          in: query
          description: '"fleet" (default) or "target:{key}".'
          schema:
            type: string
        - name: t
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Triggering states in the bucket
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TriggeringState'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /state/doc:
    post:
      summary: Ingest a statekit display document
      description: Accepts the YAML document served by a component's /state endpoint.
      parameters:
        - name: observed_at
          in: query
          required: false
          schema:
            type: string
            format: date-time
      requestBody:
        required: true
        content:
          text/yaml:
            schema:
              $ref: '#/components/schemas/StateDisplayDocument'
      responses:
        '204':
          description: Stored
  /state/mutes:
    get:
      summary: List active state mutes
      responses:
        '200':
          description: Active mutes
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateMute'
        '304':
          $ref: '#/components/responses/NotModified'
    post:
      summary: Create or update a state mute
      description: >
        Caps one state's effective status until expiration. Use status pass to
        force-pass the state. Either duration (Go duration such as 30m, plus a
        day suffix such as 7d) or expires_at must be provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MuteRequest'
      responses:
        '200':
          description: Stored mute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateMute'
        '400':
          $ref: '#/components/responses/BadRequest'
  /state/mutes/{identity}:
    delete:
      summary: Clear a state mute
      parameters:
        - name: identity
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cleared
  /escalations/doc:
    post:
      summary: Ingest an escalation display document
      description: Accepts the YAML document served by a component's /escalations endpoint.
      parameters:
        - name: source
          in: query
          required: true
          schema:
            type: string
        - name: observed_at
          in: query
          required: false
          schema:
            type: string
            format: date-time
      requestBody:
        required: true
        content:
          text/yaml:
            schema:
              $ref: '#/components/schemas/EscalationDisplayDocument'
      responses:
        '204':
          description: Stored
  /escalations/incidents:
    get:
      summary: List stored escalation incidents (summaries)
      description: >
        Incident summaries without events. Fetch
        /escalations/incidents/{source}/{id} for the event log. YAML by
        default, JSON when the Accept header includes application/json.
      parameters:
        - name: source
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, acknowledged]
        - name: id
          in: query
          schema:
            type: string
        - name: type
          in: query
          description: Incident type, for example build, deployment, or rollback.
          schema:
            type: string
      responses:
        '200':
          description: Escalation incident summaries
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            text/yaml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'
        '304':
          $ref: '#/components/responses/NotModified'
  /escalations/incidents/{source}/{id}:
    get:
      summary: One stored escalation incident, with events
      description: >
        The full incident including its bounded event ring. When the same
        source and id were reused across restarts, the most recently created
        incident is returned.
      parameters:
        - name: source
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Escalation incident
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
  /escalations/global:
    post:
      summary: Report a fleet-wide incident
      description: >
        Creates or updates a global incident such as a build, deployment, or
        rollback, without a component owning it. Omit id to create a new
        incident; include id (and optionally status closed) to annotate or
        close an existing one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalIncidentRequest'
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '201':
          description: Created incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /escalations/ack:
    post:
      summary: Mark a stored escalation incident acknowledged
      parameters:
        - name: source
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Acknowledged
...
components:
  headers:
    ETag:
      description: Strong entity tag for conditional polling with If-None-Match.
      schema:
        type: string
  responses:
    NotModified:
      description: Content unchanged since the If-None-Match tag.
    NotFound:
      description: Subject does not exist.
      content:
        text/plain:
          schema:
            type: string
    BadRequest:
      description: Invalid parameters.
      content:
        text/plain:
          schema:
            type: string
  schemas:
    FleetSummary:
      type: object
      required: [worst_status, status_counts, targets, fleet_hash, observed_at]
      properties:
        worst_status:
          type: string
          enum: [pass, warn, fail, down]
        status_counts:
          description: Count of states by status, including checks.
          type: object
          additionalProperties:
            type: integer
        targets:
          type: object
          required: [total, by_status]
          properties:
            total:
              type: integer
            by_status:
              description: Count of targets by their worst status.
              type: object
              additionalProperties:
                type: integer
        fleet_hash:
          description: Hash of the sorted (target key, material_hash) pairs; the L1 ETag.
          type: string
        observed_at:
          type: string
          format: date-time
...
    TargetDetail:
      allOf:
        - $ref: '#/components/schemas/TargetSummary'
        - type: object
          required: [details]
          properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/StateDetail'
...
    StateTimeline:
      type: object
      required: [identity, transitions]
      properties:
        identity:
          type: string
        transitions:
          description: Newest first.
          type: array
          items:
            $ref: '#/components/schemas/Transition'
...